requirements.txt
# Google Scholar Citation Tracking - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
python-docx>=1.1.0          # For reading Word documents (.docx)
ddgs>=9.0.0                 # For web search to find Google Scholar profiles
beautifulsoup4>=4.12.0      # For HTML parsing

# Optional: Direct Google Scholar access (often gets blocked)
scholarly>=1.7.11           # For accessing Google Scholar data directly
